Create file with initial condition variables
In order to be able to acess what kind of initial condition was 
generated (i.e. prime, random, size of block and m-vector etc.) a file 
is created containing these information.

Once you shall not change the initial conditions from now on, the only 
diference between file names was the type (Prime or Random).