Optimize file and variable names
There was a problem of having two different formats of file name for the 
types "Prime" and "Random". This commit creates an unique format, 
optimizing the process, without necessity of ifs for saving and reading 
depending on "type" variable.

Variable names also were renewed, with the objective of be more clear.